Add tabindex=-1 to the permalink anchor element
We use the `markdown-it-anchor` plugin to render permalink anchor elements and a link SVG onto our header elements. We tell the plugin to add aria-hidden so that the anchors are removed from screen readers and other assitive tech. There's a WCAG rule that says any element with aria-hidden should also be removed from the list of focuasable elements. One way to remove an element from the focusability list is to add the tabindex=-1 property which is what this commit does.

For more information see: https://dequeuniversity.com/rules/axe/4.10/aria-hidden-focus?application=webdriverjs