Set header content to hidden so it doesn't temporarily show open on slower loads
I thought the USWDS-init JS that is loaded in `<head>` was supposed to handle this but I was occassionaly still seeing the unfurled gov header briefly during some loads. It's easiest to see if you emulate mid-tier mobile throttling and reload the page. Setting it to be hidden fixes this.